requirements.txt
argparse
json
pyyaml
xmltodict
